feat(loki-query-editor): replace unnecessary class with class method
