+++ public/app/plugins/datasource/loki/variables.test.ts
@@ -1,11 +1,11 @@
 import { TemplateSrv } from 'app/features/templating/template_srv';
 
-import { LokiMetricFindQuery } from './metricFindQuery';
 import { createLokiDatasource, createMetadataRequest } from './mocks';
 import { LokiVariableQueryType } from './types';
+import { LokiVariableSupport } from './variables';
 
-describe('LokiMetricFindQuery', () => {
-  let lokiMetricFindQuery: LokiMetricFindQuery;
+describe('LokiVariableSupport', () => {
+  let lokiVariableSupport: LokiVariableSupport;
 
   beforeEach(() => {
     const datasource = createLokiDatasource({} as unknown as TemplateSrv);
@@ -18,19 +18,19 @@ describe('LokiMetricFindQuery', () => {
         )
       );
 
-    lokiMetricFindQuery = new LokiMetricFindQuery(datasource);
+    lokiVariableSupport = new LokiVariableSupport(datasource);
   });
 
   it('should return label names for Loki', async () => {
     // label_names()
-    const response = await lokiMetricFindQuery.execute({ refId: 'test', type: LokiVariableQueryType.LabelNames });
+    const response = await lokiVariableSupport.execute({ refId: 'test', type: LokiVariableQueryType.LabelNames });
 
     expect(response).toEqual([{ text: 'label1' }, { text: 'label2' }]);
   });
 
   it('should return label values for Loki when no matcher', async () => {
     // label_values(label1)
-    const response = await lokiMetricFindQuery.execute({
+    const response = await lokiVariableSupport.execute({
       refId: 'test',
       type: LokiVariableQueryType.LabelValues,
       label: 'label1',
@@ -41,7 +41,7 @@ describe('LokiMetricFindQuery', () => {
 
   it('should return label values for Loki with matcher', async () => {
     // label_values({label1="value1", label2="value2"},label5)
-    const response = await lokiMetricFindQuery.execute({
+    const response = await lokiVariableSupport.execute({
       refId: 'test',
       type: LokiVariableQueryType.LabelValues,
       stream: '{label1="value1", label2="value2"}',
