+++ public/app/plugins/datasource/loki/variables.ts
@@ -1,29 +1,40 @@
 import { from, Observable } from 'rxjs';
-import { map, mergeMap } from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 
 import { CustomVariableSupport, DataQueryRequest, DataQueryResponse } from '@grafana/data';
 
 import { LokiVariableQueryEditor } from './components/VariableQueryEditor';
 import { LokiDatasource } from './datasource';
-import { LokiMetricFindQuery } from './metricFindQuery';
-import { LokiVariableQuery } from './types';
+import { LokiVariableQuery, LokiVariableQueryType } from './types';
 
 export class LokiVariableSupport extends CustomVariableSupport<LokiDatasource, LokiVariableQuery> {
-  private metricFindQuery: LokiMetricFindQuery;
   editor = LokiVariableQueryEditor;
 
-  constructor(datasource: LokiDatasource) {
+  constructor(private datasource: LokiDatasource) {
     super();
-    this.metricFindQuery = new LokiMetricFindQuery(datasource);
     this.query = this.query.bind(this);
   }
 
-  query(request: DataQueryRequest<LokiVariableQuery>): Observable<DataQueryResponse> {
-    const executeObservable = this.metricFindQuery.execute(request.targets[0]);
+  async execute(query: LokiVariableQuery) {
+    if (query.type === LokiVariableQueryType.LabelNames) {
+      return this.datasource.labelNamesQuery();
+    }
 
-    return from(executeObservable).pipe(
-      mergeMap(() => executeObservable),
-      map((data) => ({ data }))
-    );
+    if (!query.label) {
+      return [];
+    }
+
+    // If we have query expr, use /series endpoint
+    if (query.stream) {
+      return this.datasource.labelValuesSeriesQuery(query.stream, query.label);
+    }
+
+    return this.datasource.labelValuesQuery(query.label);
+  }
+
+  query(request: DataQueryRequest<LokiVariableQuery>): Observable<DataQueryResponse> {
+    const result = this.execute(request.targets[0]);
+
+    return from(result).pipe(map((data) => ({ data })));
   }
 }
